+++ pkg/proxy/ipvs/graceful_termination_test.go
@@ -26,8 +26,8 @@ import (
 
 	netutils "k8s.io/utils/net"
 
-	utilipvs "k8s.io/kubernetes/pkg/util/ipvs"
-	utilipvstest "k8s.io/kubernetes/pkg/util/ipvs/testing"
+	utilipvs "k8s.io/kubernetes/pkg/proxy/ipvs/util"
+	utilipvstest "k8s.io/kubernetes/pkg/proxy/ipvs/util/testing"
 )
 
 func Test_GracefulDeleteRS(t *testing.T) {
