+++ pkg/proxy/ipvs/util/testing/fake_test.go
@@ -19,7 +19,7 @@ package testing
 import (
 	"testing"
 
-	utilipvs "k8s.io/kubernetes/pkg/util/ipvs"
+	utilipvs "k8s.io/kubernetes/pkg/proxy/ipvs/util"
 	netutils "k8s.io/utils/net"
 )
