--- pkg/util/ipvs/OWNERS
@@ -1,15 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-reviewers:
-  - sig-network-reviewers
-  - aroradaman
-  - uablrek
-approvers:
-  - sig-network-approvers
-  - uablrek
-labels:
-  - sig/network
-  - area/ipvs
-emeritus_approvers:
-  - lbernail
-  - m1093782566
